dstore: Implement abstract layer
(cherry-picked from 8038cbf00b28137df1d0ad16a7444e0c68233329)
